+++ collisionevent.h
@@ -1,9 +1,10 @@
 #ifndef COLLISIONEVENT_H
 #define COLLISIONEVENT_H
+#include <glm/glm.hpp>
 
 class CollisionEvent {
   public:
-    CollisionEvent(int object, float time,
+    CollisionEvent(int object, double time,
                     glm::vec3 initial_coordinates_, glm::vec3 initial_axis,
                     float initial_angle,
                     glm::vec3 axis_of_rotation_, glm::vec3 velocity,
